refactor(ui): extract ConnectLogsSection into its own component
Move the log table out of Connect.jsx and into components/ConnectLogsSection.jsx, alongside SystemEvents.jsx, per PR #1333 review feedback.
